Unify text insertion when parsing HTML and XML
