+++ tests/wpt/metadata/dom/nodes/ProcessingInstruction-literal-1.xhtml.ini
@@ -1,5 +1,5 @@
 [ProcessingInstruction-literal-1.xhtml]
   type: testharness
-  [<]
+  [<?xml?> is not a ProcessingInstruction]
     expected: FAIL
